requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
jinja2
python-multipart
bcrypt
python-docx
itsdangerous
apscheduler
